backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
cryptography==44.0.2
cymem==2.0.11
daphne==4.1.2
Django==5.1.6
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
filelock==3.17.0
fsspec==2025.2.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
huggingface-hub==0.29.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
joblib==1.4.2
kombu==5.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
murmurhash==1.0.12
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
preshed==3.0.9
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
Twisted==24.11.0
txaio==23.1.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
zope.interface==7.2
